% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf_scraper2.R
\name{a_pdf_scraper_2}
\alias{a_pdf_scraper_2}
\title{A pdf scraper function v2
This function allows you to scrap specific pdfs.
which prints 'Hello, world!'.
Example:
a_pdf_scraper(pdf_location, csv_location)}
\usage{
a_pdf_scraper_2(pdf_location, csv_location)
}
\description{
A pdf scraper function v2
This function allows you to scrap specific pdfs.
which prints 'Hello, world!'.
Example:
a_pdf_scraper(pdf_location, csv_location)
}
